delete cluster, rename rmon_tools
