+++ resource_monitor/src/resource_monitor_tools.h
